[TS-356] Refactor: ChattingMEssage 파일 옮김
